Add native Windows (MSVC + vcpkg) support
Port the Vulkan build to run on a native Windows driver alongside the
existing macOS/MoltenVK path, from a single code path:

- createInstance: feature-detect portability extensions instead of
  hard-coding the macOS surface/portability extensions (which crashed at
  instance creation on a native driver)
- pickPhysicalDevice: prefer a discrete GPU and log the choice, so
  laptops with switchable graphics don't run the sim on the integrated GPU
- OpenMP: signed loop indices and an indexed (not range-based) loop, as
  required by MSVC's OpenMP 2.0
- main.cpp: include <print> explicitly (MSVC does not pull it in transitively)
- GridRenderer: NOMINMAX + #undef MemoryBarrier so <windows.h> does not
  clobber std::max / vk::MemoryBarrier
- CMake: find_package(glm); vcpkg.json manifest installs GLFW + GLM
- .gitignore: build dirs, vcpkg_installed, machine-specific CMakeUserPresets.json
- README: cross-platform build docs (Windows/vcpkg flow + GPU selection)

+++ src/Geometry.cpp
@@ -173,7 +173,7 @@ void FlatGeometry::warpGrid(
     const std::vector<Vertex>& base = baseGrid();
 
 #pragma omp parallel for
-    for ( size_t v = 0; v < base.size(); ++v )
+    for ( int v = 0; v < static_cast<int>(base.size()); ++v )
     {
         Vertex vertex = base[v];
         const float rawDepth = warpDepth(vertex.position, bodies, warp);
@@ -348,7 +348,7 @@ void SphericalGeometry::warpGrid(
     const std::vector<Vertex>& base = baseGrid();
 
 #pragma omp parallel for
-    for ( size_t v = 0; v < base.size(); ++v )
+    for ( int v = 0; v < static_cast<int>(base.size()); ++v )
     {
         Vertex vertex = base[v];
 
@@ -505,7 +505,7 @@ void HyperbolicGeometry::warpGrid(
     const float k = m_GridScale; // paraboloid scale: y = (x^2 - z^2) / k
 
 #pragma omp parallel for
-    for ( size_t v = 0; v < base.size(); ++v )
+    for ( int v = 0; v < static_cast<int>(base.size()); ++v )
     {
         Vertex vertex = base[v];
         const float rawDepth = warpDepth(vertex.position, bodies, warp);
